requirements.txt
aiogram>=3.20
aiohttp>=3.9
python-dotenv>=1.0
